Add release notes filters for version 4.0.4 (#177)
* Add release notes filters for version 4.0.4

Introduced acp_fixed_4.0.4 and acp_known_issues_4.0.4 filters to support release notes and known issues tracking for version 4.0.4 in doom.config.yml.

* Add release notes section for 4.0.4 and update filters

Introduced a new 4.0.4 section in the release notes and updated the corresponding filters in doom.config.yml for fixed and known issues to target version 4.0.4. This ensures accurate tracking and display of issues relevant to the 4.0.4 release.

---------

Co-authored-by: MeoK <12069138+chinameok@users.noreply.github.com>